💄Remove card buttons from eating modal

+++ src/lib/components/atoms/ItemCard.svelte
@@ -76,7 +76,7 @@
 				/>
 			</svg>
 		</button>
-	{:else}
+	{:else if type === 'bright' && !eatingMenu}
 		<!-- Removing Item -->
 		<button class="absolute right-1 top-0.5" on:click={() => removeFromPlannedItems(id)}>
 			<svg class="h-3 fill-neutral-200/50" viewBox="0 0 24 24">
@@ -88,16 +88,18 @@
 		</button>
 	{/if}
 	<!-- Image Icon: Edit -->
-	<button class="absolute bottom-[calc(40px)] left-[calc(2px)]" on:click={() => deleteItem(id)}>
-		<svg class="h-4 w-4 fill-none stroke-neutral-200/70" viewBox="0 0 24 24">
-			<path
-				stroke-linecap="round"
-				stroke-linejoin="round"
-				stroke-width="1"
-				d="M12 4H6a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-6M9 15v-2.5l8.75-8.75c.69-.69 1.81-.69 2.5 0v0c.69.69.69 1.81 0 2.5L15.5 11l-4 4H9Z"
-			/>
-		</svg>
-	</button>
+	{#if !eatingMenu}
+		<button class="absolute bottom-[calc(40px)] left-[calc(2px)]" on:click={() => deleteItem(id)}>
+			<svg class="h-4 w-4 fill-none stroke-neutral-200/70" viewBox="0 0 24 24">
+				<path
+					stroke-linecap="round"
+					stroke-linejoin="round"
+					stroke-width="1"
+					d="M12 4H6a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2v-6M9 15v-2.5l8.75-8.75c.69-.69 1.81-.69 2.5 0v0c.69.69.69 1.81 0 2.5L15.5 11l-4 4H9Z"
+				/>
+			</svg>
+		</button>
+	{/if}
 	<!-- Item Name -->
 	<p
 		class="truncate pl-1 pt-1 text-[9px] font-medium {type === 'dark'
